[IMP] pos_restaurant: improve UI of order button
Before this commit when clicking on the order button in the
pos_restaurant, the order button was not indicating that a request was
being processed.

This commit adds a spinner to the order button to indicate that a
request is being processed. The spinner is displayed when the order
button is clicked and is hidden when the order is successfully
processed.
