+++ addons/pos_restaurant/static/src/overrides/components/product_screen/actionpad_widget/actionpad_widget.xml
@@ -5,9 +5,12 @@
         <xpath expr="//button[hasclass('validation')]" position="after">
             <button
                 t-if="this.swapButton"
-                t-attf-class="submit-order h-100 {{getMainButtonClasses()}}"
+                t-attf-class="submit-order h-100 {{getMainButtonClasses()}} position-relative"
                 t-att-class="swapButtonClasses"
                 t-on-click="submitOrder">
+                <div t-if="this.uiState.clicked and displayCategoryCount.length" class="h-100 w-100 position-absolute top-0 start-0 d-flex align-items-center justify-content-center" style="backdrop-filter: blur(7px);">
+                    <img src="/web/static/img/spin.svg" style="height:35px; width:35px" alt="Loading..."/>
+                </div>
                 <i class="fa fa-cutlery"></i>
                 Order
                 <div t-if="displayCategoryCount.length" class="break-line fw-normal fs-6 w-100  border-top mt-1 pt-1">
