+++ addons/pos_restaurant/static/src/overrides/components/product_screen/product_screen.js
@@ -2,7 +2,7 @@
 
 import { ProductScreen } from "@point_of_sale/app/screens/product_screen/product_screen";
 import { patch } from "@web/core/utils/patch";
-import { onMounted } from "@odoo/owl";
+import { onMounted, useState } from "@odoo/owl";
 
 patch(ProductScreen.prototype, {
     /**
@@ -10,6 +10,10 @@ patch(ProductScreen.prototype, {
      */
     setup() {
         super.setup(...arguments);
+
+        this.uiState = useState({
+            clicked: false,
+        });
 
         onMounted(() => {
             this.pos.addPendingOrder([this.pos.get_order().id]);
@@ -41,8 +45,18 @@ patch(ProductScreen.prototype, {
     get swapButton() {
         return this.pos.config.module_pos_restaurant && this.pos.orderPreparationCategories.size;
     },
-    submitOrder() {
-        this.pos.sendOrderInPreparationUpdateLastChange(this.pos.get_order());
+    get displayCategoryCount() {
+        return this.pos.categoryCount.slice(0, 3);
+    },
+    async submitOrder() {
+        if (!this.uiState.clicked) {
+            this.uiState.clicked = true;
+            try {
+                await this.pos.sendOrderInPreparationUpdateLastChange(this.currentOrder);
+            } finally {
+                this.uiState.clicked = false;
+            }
+        }
     },
     get primaryReviewButton() {
         return (
